backend/pyproject.toml
[project]
name = "inspi-backend"
version = "0.1.0"
description = "Inspi – Inspirator für Pfadfinder-Gruppenstunden (Backend)"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1,<6.0",
    "django-ninja>=1.3,<2.0",
    "django-allauth>=65.0,<66.0",
    "django-cors-headers>=4.4,<5.0",
    "django-storages[google]>=1.14,<2.0",
    "django-environ>=0.11,<1.0",
    "django-imagekit>=5.0,<6.0",
    "django-ckeditor>=6.7,<7.0",
    "django-cleanup>=9.0,<10.0",
    "psycopg[binary]>=3.2,<4.0",
    "gunicorn>=23.0,<24.0",
    "Pillow>=11.0,<12.0",
    "google-cloud-storage>=2.18,<3.0",
    "google-cloud-secret-manager>=2.22,<3.0",
    "google-genai>=1.0,<2.0",
    "pydantic>=2.10,<3.0",
    "html2text>=2025.4.15",
    "openpyxl>=3.1,<4.0",
    "reportlab>=4.2,<5.0",
    "channels>=4.0,<5.0",
    "channels-redis>=4.2,<5.0",
    "daphne>=4.1,<5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9.0",
    "pytest-django>=4.9,<5.0",
    "pytest-cov>=5.0,<6.0",
    "pytest-asyncio>=0.24,<1.0",
    "ruff>=0.8,<1.0",
    "black>=24.10,<25.0",
    "mypy>=1.13,<2.0",
    "django-stubs>=5.1,<6.0",
    "pre-commit>=4.0,<5.0",
    "django-debug-toolbar>=4.4,<5.0",
    "factory-boy>=3.3,<4.0",
    "ipython>=8.29,<9.0",
    "model-bakery>=1.19,<2.0",
    "httpx>=0.28,<1.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["idea*", "planner*", "core*", "inspi*", "packinglist*", "content*", "supply*", "session*", "blog*", "game*"]

[tool.ruff]
target-version = "py313"
line-length = 120
src = ["."]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "W",     # pycodestyle warnings
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "S",     # flake8-bandit (security)
    "T20",   # flake8-print
    "SIM",   # flake8-simplify
    "RUF",   # ruff-specific rules
]
ignore = [
    "S101",  # allow assert in tests
    "S105",  # allow hardcoded passwords in settings (dev defaults)
]

[tool.ruff.lint.per-file-ignores]
"**/tests/**" = ["S101", "S106"]
"**/settings/**" = ["S105"]
"**/migrations/**" = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["idea", "planner", "core", "inspi", "packinglist", "content", "supply", "session", "blog", "game"]

[tool.black]
line-length = 120
target-version = ["py313"]

[tool.mypy]
python_version = "3.13"
plugins = ["mypy_django_plugin.main"]
strict = false
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_defs = false

[tool.django-stubs]
django_settings_module = "inspi.settings.local"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "inspi.settings.test"
python_files = ["tests.py", "test_*.py"]
addopts = [
    "--strict-markers",
    "--tb=short",
    "-q",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.coverage.run]
source = ["idea", "planner", "core", "packinglist", "content", "supply", "session", "blog", "game"]
omit = ["*/migrations/*", "*/tests/*"]

[tool.coverage.report]
show_missing = true
fail_under = 50
